evals/tools/itag/sdk/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alibabacloud-credentials==0.3.1
alibabacloud-darabonba-array==0.1.0
alibabacloud-darabonba-number==0.0.4
alibabacloud-darabonba-string==0.0.4
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.2.1
alibabacloud-openitag20220616==1.0.1
alibabacloud-tea==0.2.9
alibabacloud-tea-fileform==0.0.5
alibabacloud-tea-openapi==0.3.6
alibabacloud-tea-util==0.3.8
alibabacloud-tea-xml==0.0.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
darabonba-rsa==0.0.7
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.17
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
urllib3==1.26.14
yarl==1.8.2
